% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convenience.R
\name{adsparse_to_admatrix}
\alias{adsparse_to_admatrix}
\title{Convert a sparse matrix to a dense matrix, preserving RTMB evaluation mode}
\usage{
adsparse_to_admatrix(m)
}
\arguments{
\item{m}{A sparse matrix}
}
\value{
A dense matrix with the appropriate evaluation class
}
\description{
Convert a sparse matrix to a dense matrix, preserving RTMB evaluation mode
}
