% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annodes.R
\name{annodes}
\alias{annodes}
\title{Find a set of nodes to cover a dataset}
\usage{
annodes(
  data,
  n_nodes = 10,
  max.it = 100,
  jitter_magnitude = 1e-04,
  create_graph = TRUE,
  k = 4
)
}
\arguments{
\item{data}{A matrix of data points. Each row should be a data point.}

\item{n_nodes}{The number of nodes to cover the dataset.}

\item{max.it}{The number of iterations to use when finding the best set of nodes.}

\item{jitter_magnitude}{The relative step size for randomly sampled jitters}

\item{create_graph}{Should a directed acyclic graph between the nodes be returned?}

\item{k}{If create_graph is TRUE, the number of parents for each node.}
}
\value{
A list giving the nodes and the objective function trajectory.
}
\description{
Find a set of nodes to cover a dataset
}
