% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lcspline.R
\name{cache_lc}
\alias{cache_lc}
\title{Precompute spline structure}
\usage{
cache_lc(spline, stretch_sequence)
}
\arguments{
\item{spline}{A lcspline object.}

\item{stretch_sequence}{A numeric vector giving the range parameters used to
precompute the spline structure.}
}
\value{
A copy of spline with an updated precomputed spline structure
}
\description{
Precompute spline structure
}
