% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convenience.R
\name{conditional_normal}
\alias{conditional_normal}
\title{Conditional mean and variance from joint Gaussian variables}
\usage{
conditional_normal(
  x,
  joint_mean,
  joint_covariance,
  predicted = 1,
  predictors = tail(seq(nrow(x)), -1)
)
}
\arguments{
\item{x}{An n x k matrix. Columns index independent replicates.}

\item{joint_mean}{An n x k mean matrix. Columns index independent replicates.}

\item{joint_covariance}{An n x n covariance matrix.}

\item{predicted}{Indices declaring which of the n elements should be predicted.}

\item{predictors}{Indices declaring which of the n elements should be conditioned on.}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{A list with the conditional mean of variance of 
x\[predicted, \] given x\[predictors, \]
}\if{html}{\out{</div>}}
}
\description{
Conditional mean and variance from joint Gaussian variables
}
