% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lcspline.R
\name{create_lcspline}
\alias{create_lcspline}
\title{Create a lcspline object}
\usage{
create_lcspline(
  x,
  n_parents = 4,
  correlation_function = function(x1, x2, range) {
     d <- sqrt(sum((x1 - x2)^2))
    
    if (range == 0) 
         return((d == 0) * 1)
     (1 + d/range) * exp(-d/range)
 },
  lc_sequence = 0,
  graph
)
}
\arguments{
\item{x}{A numeric matrix giving the locations (as rows) for which the spline
values are desired}

\item{n_parents}{The number of parents each node should have.}

\item{lc_sequence}{A numeric vector giving a sequence of range parameters used to precompute
the spline structure.}

\item{graph}{An optional argument. If supplied, it needs to be an igraph object
representing a directed acyclic graph.}

\item{covariance_function}{A function used to compute covariances between spline inputs.
Should have the arguments x1 and x2, vectors giving the spline inputs,
and a range parameter.}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{A lcspline object as a list with the following elements:
}\if{html}{\out{</div>}}
\itemize{
\item values
The output values of the spline at x.
\item map
Mapping indices for the node values. See ?TMB::MakeADFun.
\item x
The locations at which the spline is evaluated.
\item graph
An igraph directed acyclic graph represnting the node parent structure.
\item n_parents
The number of parents for each location
\item LC
A list defining the precomputed spline structure for each range value
in lc_sequence. Each element of LC is a list with length equal to
length(values). Each sub-element gives the conditional mean of the
spline value as a linear combination "plc" of its parents, and
the conditional standard deviation as a scalar "elc".
\item lc_sequence
The sequence of range parameters used to compute LC.
\item mixtape
An AD-compatible function used to compute mixture probabilities used
to interpolate between precomputed values of LC.
\item correlation_function
The function used to compute correlations between input points.
}
}
\description{
Create a lcspline object
}
