% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spline.R
\name{create_nnspline}
\alias{create_nnspline}
\title{Create a nnspline object}
\usage{
create_nnspline(
  x,
  nodes = x,
  n_parents = 4,
  variance = 1,
  parameters = c(1),
  correlation_function = function(x1, x2, p) {
     d <- sqrt(sum((x1 - x2)^2))
    
    range <- p[[1]]
     (1 + d/range) * exp(-d/range)
 }
)
}
\arguments{
\item{x}{A numeric matrix giving the locations (as rows) for which the spline values are desired}

\item{nodes}{A numeric matrix giving the spline knot locations as rows.}

\item{n_parents}{The number of parents each node should have.}

\item{variance}{The marginal variance of the spline.}

\item{parameters}{The parameters used in the correlation function.}

\item{correlation_function}{A function used to compute correlation between spline inputs. Should have the arguments x1 and x2, vectors giving the spline inputs, and p, a vector of parameters.}
}
\value{
A multispline object as a list with the following elements:
\itemize{
\item values The output values of the spline at x.
\item x The locations at which the spline is evaluated.
\item x_parents The nodes which are parents of each x.
\item node_values The output values of the spline at the nodes.
\item node_values_map Mapping indices for the node values. See ?TMB::MakeADFun.
\item nodes The locations of the spline nodes.
\item node_graph An igraph directed acyclic graph represnting the node parent structure.
\item variance The marginal variance of the spline.
\item parameters The parameters for the correlation function.
\item parameter_map Mapping indices for the parameters. See ?TMB::MakeADFun.
\item correlation_function The function used to compute correlations between input points.
\item x_covariance Sparse covariance matrices for the spline x values.
\item node_covariance Sparse covariance matrices for the spline nodes.
}
}
\description{
Create a nnspline object
}
