% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lcspline.R
\name{dlcspline}
\alias{dlcspline}
\title{Evaluate the log-likelihood of a lcspline}
\usage{
dlcspline(x, spline, stretch, height, log = TRUE)
}
\arguments{
\item{x}{The output values of the spline.}

\item{spline}{A lcspline object.}

\item{stretch}{A numeric value between 0 and 1 determining the stretch of the spline.
The value between 0 and 1 is shifted and scaled to the minimum and
maximum values of the spline's stretch_sequence.}

\item{height}{A positive numeric value proportional to the height of the spline. The
conditional variance is multiplied by height^2.}

\item{log}{If true, returns the log-likelihood.}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{The log-likelihood of the spline evaluated at x.
}\if{html}{\out{</div>}}
}
\description{
Evaluate the log-likelihood of a lcspline
}
