% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spline.R
\name{dpspline}
\alias{dpspline}
\title{Evaluate the penalized log-likelihood of a spline}
\usage{
dpspline(x, spline, penalty = 0.01, log = TRUE)
}
\arguments{
\item{x}{The node values}

\item{spline}{The spline object}

\item{penalty}{How much should the variance be penalized?}

\item{log}{If TRUE, returns the log-likelihood.}
}
\value{
The log-likelihood of the spline evaluated at x. If x is a simref
object, then the values of x will be replaced with simulated values.
}
\description{
Evaluate the penalized log-likelihood of a spline
}
