% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spline.R
\name{dspline}
\alias{dspline}
\title{Evaluate the log-likelihood of a spline}
\usage{
dspline(x, spline, log = TRUE)
}
\arguments{
\item{x}{The node values}

\item{spline}{The spline object}

\item{log}{If TRUE, returns the log-likelihood.}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{The log-likelihood of the spline evaluated at x. If x is a simref
object, then the values of x will be replaced with simulated values.
}\if{html}{\out{</div>}}
}
\description{
Evaluate the log-likelihood of a spline
}
