% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lcspline.R
\name{find_upper_stretch}
\alias{find_upper_stretch}
\title{Find a sensible upper limit for the range parameter}
\usage{
find_upper_stretch(x, graph, correlation_function, target_correlation = 0.9)
}
\arguments{
\item{x}{The spline x coordinates}

\item{graph}{The spline graph}

\item{correlation_function}{The correlation function}

\item{target_correlation}{The target value for the correlation of the
furthest apart neighbor coordinates at the the upper limit of the range
parameter.}
}
\value{
A numeric value giving a sensible upper limit for the range parameter.
}
\description{
This function makes a few assumptions about the spline correlation function.
1.) Stationariy: it depends only on the distance between two points
and not the actual location of those points.
2.) Monotone decreasing in distance.
3.) Approaches 0 as distance goes to infinity.
It finds the range parameter at which the correlation between any x and any
of its parents is at least a target value.
}
