% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convenience.R
\name{get_joint_covariance}
\alias{get_joint_covariance}
\title{Find the joint covariance matrix for an x/node and its parents}
\usage{
get_joint_covariance(
  spline,
  idx,
  parent_idx,
  prediction_type = "node",
  mode = "numeric"
)
}
\arguments{
\item{spline}{The spline}

\item{idx}{The index of either an x or node}

\item{parent_idx}{The index of the parents of idx}

\item{prediction_type}{Either "x" if idx refers to an x or "node" if idx refers to a node}

\item{mode}{Either "numeric" or "advector"}
}
\value{
A dense covariance matrix
}
\description{
Find the joint covariance matrix for an x/node and its parents
}
