% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convenience.R
\name{get_pairs}
\alias{get_pairs}
\title{Find each pair of nodes whose correlation needs to be computed.}
\usage{
get_pairs(dag)
}
\arguments{
\item{dag}{An directed acyclic graph as an igraph object}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{A data.frame giving the list of every node that appears in the same
parent list or are parent/children.
}\if{html}{\out{</div>}}
}
\description{
Find each pair of nodes whose correlation needs to be computed.
}
