% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convenience.R
\name{ip_to_j}
\alias{ip_to_j}
\title{Convert an (i, p) sparse matrix index to an (i, j) index}
\usage{
ip_to_j(i, p)
}
\arguments{
\item{i}{the i index}

\item{p}{the p index}
}
\value{
A vector giving the column index of the entries
}
\description{
Convert an (i, p) sparse matrix index to an (i, j) index
}
