% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nnodes.R
\name{nnodes}
\alias{nnodes}
\title{Find a good set of nodes to cover a dataset}
\usage{
nnodes(
  data,
  n_nodes = 10,
  max.it = 100,
  jitter_magnitude = 1e-04,
  create_graph = FALSE,
  n_parents = 4
)
}
\arguments{
\item{data}{A matrix of data points. Each row should be a data point.}

\item{n_nodes}{The number of nodes to cover the dataset.}

\item{max.it}{The number of iterations to use when finding the best set of nodes.}

\item{jitter_magnitude}{The relative step size for randomly sampled jitters}

\item{create_graph}{Should a directed acyclic graph between the nodes be returned?}

\item{n_parents}{If create_graph is TRUE, the number of parents for each node.}
}
\value{
A list giving the nodes, an LT matrix for computing mahalanobis distance, the objective function trajectory, and possibly the node graph.
}
\description{
Find a good set of nodes to cover a dataset
}
