% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lcspline.R
\name{rlcspline}
\alias{rlcspline}
\title{Simulated a lcspline}
\usage{
rlcspline(n = 1, spline, stretch, height, center = FALSE)
}
\arguments{
\item{n}{The number of simulations to return.}

\item{spline}{The lcspline to simulate from.}

\item{stretch}{A numeric value between 0 and 1 determining the stretch of the spline.
The value between 0 and 1 is shifted and scaled to the minimum and
maximum values of the spline's stretch_sequence.}

\item{height}{A positive numeric value proportional to the height of the spline. The
conditional variance is multiplied by height^2.}

\item{center}{Logical. If true, the simulated spline values will have mean 0.}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{Values simulated from the spline.
}\if{html}{\out{</div>}}
}
\description{
Simulated a lcspline
}
