% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nnspline.R
\name{rspline}
\alias{rspline}
\title{Simulate a spline}
\usage{
rspline(spline, n = 1, center = FALSE)
}
\arguments{
\item{spline}{The spline to simulate from}

\item{n}{The number of samples}

\item{center}{Logical. If true, the spline values will have mean 0.}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{A list of length n with the simulated splines. If n == 1 then the 
spline will be returned not in a list.
}\if{html}{\out{</div>}}
}
\description{
Simulate a spline
}
