% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nnodes.R
\name{symsqrt}
\alias{symsqrt}
\title{Compute symmetric square roots of positive definite matrices}
\usage{
symsqrt(m, invert = FALSE)
}
\arguments{
\item{m}{A positive definite matrix}

\item{invert}{If TRUE, the take the symmetric square root of m^-1}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{A symmetric matrix x where x \%*\% x == m (or m^-1)
}\if{html}{\out{</div>}}
}
\description{
Compute symmetric square roots of positive definite matrices
}
