% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lcspline.R
\name{cache_lc}
\alias{cache_lc}
\title{Precompute spline structure}
\usage{
cache_lc(lc_sequence, spline)
}
\arguments{
\item{lc_sequence}{A numeric vector giving the range parameters used to
precompute the spline structure.}

\item{spline}{A lcspline object.}
}
\value{
A copy of spline with an updated precomputed spline structure
}
\description{
Precompute spline structure
}
