% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convenience.R
\name{conditional_normal}
\alias{conditional_normal}
\title{Conditional mean and variance from joint Gaussian variables}
\usage{
conditional_normal(
  x,
  mu,
  Sigma,
  predicted = 1,
  predictors = tail(seq(nrow(x)), -1)
)
}
\arguments{
\item{x}{An n x k matrix. Columns index independent replicates.}

\item{mu}{An n x k mean matrix. Columns index independent replicates.}

\item{Sigma}{An n x n covariance matrix. Can be a sparse matrix.}

\item{predicted}{Indices declaring which of the n elements should be predicted.}

\item{predictors}{Indices declaring which of the n elements should be conditioned on.}
}
\value{
A list with the conditional mean of variance of x[predicted, ] given x[predictors, ]
}
\description{
Conditional mean and variance from joint Gaussian variables
}
