% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spline.R
\name{create_nnspline}
\alias{create_nnspline}
\title{Create a nnspline object}
\usage{
create_nnspline(
  x,
  nodes = x,
  n_parents = 4,
  parameters = c(1, 1),
  noise = 0,
  covariance_function = function(x1, x2, p) {
     d <- sqrt(sum((x1 - x2)^2))
    
    variance <- p[[1]]
     range <- p[[2]]
     cov <- variance * (1 + d/range) *
    exp(-d/range)
     return(cov)
 },
  node_graph
)
}
\arguments{
\item{x}{A numeric matrix giving the locations (as rows) for which the spline
values are desired}

\item{nodes}{A numeric matrix giving the spline knot locations as rows.}

\item{n_parents}{The number of parents each node should have.}

\item{parameters}{The parameters used in the covariance function.}

\item{noise}{A non-negative variance parameter used to model uncorrelated noise.}

\item{covariance_function}{A function used to compute covariances between spline inputs.
Should have the arguments x1 and x2, vectors giving the spline inputs,
and p, a vector of parameters.}

\item{node_graph}{An optional argument. If supplied, it needs to be an igraph object
representing a directed acyclic graph.}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{A multispline object as a list with the following elements:
}\if{html}{\out{</div>}}
\itemize{
\item values
The output values of the spline at x.
\item x
The locations at which the spline is evaluated.
\item x_parents
The nodes which are parents of each x. If a parent is a negative value
then the node value of that parent is taken to be the corresponding
x value.
\item node_values
The output values of the spline at the nodes.
\item node_values_map
Mapping indices for the node values. See ?TMB::MakeADFun.
\item nodes
The locations of the spline nodes.
\item node_graph
An igraph directed acyclic graph represnting the node parent structure.
\item n_parents
The number of parents for each location
\item parameters
The parameters for the covariance function.
\item parameter_map
Mapping indices for the parameters. See ?TMB::MakeADFun.
\item noise
A non-negative variance parameter used to model uncorrelated noise.
\item covariance_function
The function used to compute covariances between input points.
It should take argument d1 and d2, each of which represents a row
of either spline$x or spline$nodes, and an argument p giving
the parameters of the covariance function.
\item x_covariance
A list giving the covariance between x its parents. The first element
of each is the covariance of x with itself.
\item node_covariance
A partially filled in covariance matrix for the spline nodes. Only the
elements that are needed to either compute the spline likelihood or
update the spline values are filled in, all other entries are set
equal to -Inf.
\item node_pairs
A matrix giving the indices of node_covariance that need to be
filled in.
}
}
\description{
Create a nnspline object
}
