% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nnspline.R
\name{create_nnspline}
\alias{create_nnspline}
\title{Create a nnspline object}
\usage{
create_nnspline(
  x,
  n_parents = 4,
  parameters = c(1, 1),
  noise = 0,
  covariance_function = function(x1, x2, p) {
     d <- sqrt(sum((x1 - x2)^2))
    
    variance <- p[[1]] * p[[2]]^(-3)
     range <- p[[2]]
     cov <- variance * (1 +
    d/range) * exp(-d/range)
     return(cov)
 },
  graph
)
}
\arguments{
\item{x}{A numeric matrix giving the locations (as rows) for which the spline
values are desired}

\item{n_parents}{The number of parents each node should have.}

\item{parameters}{The parameters used in the covariance function.}

\item{noise}{A non-negative variance parameter used to model uncorrelated noise.}

\item{covariance_function}{A function used to compute covariances between spline inputs.
Should have the arguments x1 and x2, vectors giving the spline inputs,
and p, a vector of parameters.}

\item{graph}{An optional argument. If supplied, it needs to be an igraph object
representing a directed acyclic graph.}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{A spline object as a list with the following elements:
}\if{html}{\out{</div>}}
\itemize{
\item values
The output values of the spline at x.
\item x
The locations at which the spline is evaluated.
\item map
Mapping indices for the node values. See ?TMB::MakeADFun.
\item graph
An igraph directed acyclic graph represnting the node parent structure.
\item n_parents
The number of parents for each location
\item parameters
The parameters for the covariance function.
\item parameter_map
Mapping indices for the parameters. See ?TMB::MakeADFun.
\item noise
A non-negative variance parameter used to model uncorrelated noise.
\item covariance_function
The function used to compute covariances between input points.
It should take argument d1 and d2, each of which represents a row
of either spline$x or spline$nodes, and an argument p giving
the parameters of the covariance function.
\item covariance_matrix
A partially filled in covariance matrix for the spline nodes. Only the
elements that are needed to either compute the spline likelihood or
update the spline values are filled in, all other entries are set
equal to -Inf.
\item pairs
A matrix giving the indices of covariance_matrix that need to be
filled in.
}
}
\description{
Create a nnspline object
}
