% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convenience.R
\name{distance_matrix_to_dag}
\alias{distance_matrix_to_dag}
\title{Convert a distance matrix to a directed acyclic graph}
\usage{
distance_matrix_to_dag(d, k = 4)
}
\arguments{
\item{d}{A distance matrix, or of class "dist"new_}

\item{k}{The number of parents for each node}
}
\value{
A directed acyclic graph from the igraph package
}
\description{
Convert a distance matrix to a directed acyclic graph
}
