% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convenience.R
\name{get_pairs}
\alias{get_pairs}
\title{Find each pair of nodes whose correlation needs to be computed.}
\usage{
get_pairs(dag, x_parents)
}
\arguments{
\item{dag}{An directed acyclic graph as an igraph object}

\item{x_parents}{A sparse matrix which has a 1 in entry (i, j) if
node i is a parent of x[j]}
}
\value{
A data.frame giving the list of every node
that appear in the same parent list or are parent/children.
}
\description{
Find each pair of nodes whose correlation needs to be computed.
}
