% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spline.R
\name{nns}
\alias{nns}
\title{Evaluate a nnspline}
\usage{
nns(x, spline)
}
\arguments{
\item{x}{The input values}

\item{spline}{The spline}
}
\value{
The value(s) of the spline at x
}
\description{
Evaluate a nnspline
}
