% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lcspline.R
\name{rlcspline}
\alias{rlcspline}
\title{Simulated a lcspline}
\usage{
rlcspline(spline, smoothness, height)
}
\arguments{
\item{spline}{The lcspline to simulate from}

\item{smoothness}{A numeric value between 0 and 1 determining the smoothness of the spline.}

\item{height}{A positive numeric value proportional to the heigh of the spline.}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{A copy of the spline with new simulated values.
}\if{html}{\out{</div>}}
}
\description{
Simulated a lcspline
}
