% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spline.R
\name{rspline}
\alias{rspline}
\title{Simulate a spline}
\usage{
rspline(n = 1, spline)
}
\arguments{
\item{n}{The number of samples}

\item{spline}{The spline to simulate from}
}
\value{
A list of length n with the simulated splines
}
\description{
Simulate a spline
}
