% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spline.R
\name{rspline}
\alias{rspline}
\title{Simulate a spline}
\usage{
rspline(spline, n = 1)
}
\arguments{
\item{spline}{The spline to simulate from}

\item{n}{The number of samples}
}
\value{
A list of length n with the simulated splines. If n == 1 then the spline will be returned not in a list.
}
\description{
Simulate a spline
}
