% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nnspline.R
\name{update_spline}
\alias{update_spline}
\alias{update_parameters}
\alias{update_values}
\title{Update a spline based on node values and parameters values}
\usage{
update_spline(spline, parameters, values, ...)

update_parameters(spline, parameters, noise, ...)

update_values(spline, values, ...)
}
\arguments{
\item{spline}{An nnspline}

\item{parameters}{The new parameters for the spline. If missing, will use the pre-existing
parameters stored in the spline.}

\item{...}{Used to pass arguments onto update_parameters and
update_values.}

\item{noise}{The new variance parameter for uncorrelated noise. If missing, will use
the pre-existing noise parameter stored in the spline.}

\item{node_values}{The new new values for the spline. If missing, will use the pre-existing
node values stored in the spline.}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{An updated copy of the spline.
}\if{html}{\out{</div>}}
}
\description{
Update a spline based on node values and parameters values
}
\section{Functions}{
\itemize{
\item \code{update_parameters()}: Update a spline's covariance matrix

\item \code{update_values()}: Update a spline's values

}}
