% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spline.R
\name{update_spline_covariance}
\alias{update_spline_covariance}
\alias{update_spline_values}
\alias{update_spline}
\title{Update a spline based on node values and parameters values}
\usage{
update_spline_covariance(spline, variance, parameters, ...)

update_spline_values(spline, node_values, ...)

update_spline(spline, ...)
}
\arguments{
\item{spline}{An nnspline}

\item{variance}{The new marginal variance of the spline. If missing, will
use the pre-existing variance stored in the spline.}

\item{parameters}{The new parameters of the spline. If missing, will
use the pre-existing parameters stored in the spline.}

\item{...}{Used to pass arguments onto sub-functions.}

\item{node_values}{The new new values for the spline. If missing, will use
the pre-existing node values stored in the spline.}
}
\value{
An updated copy of the spline.
}
\description{
Update a spline based on node values and parameters values
}
\section{Functions}{
\itemize{
\item \code{update_spline_covariance()}: Update a spline's covariance matrix

\item \code{update_spline_values()}: Update a spline's values

}}
